+++ dse_do_dashboard/main_pages/prepare_data_page.py
@@ -14,43 +14,69 @@ from dse_do_dashboard.utils.dash_common_utils import get_data_table_card_childre
 
 
 class PrepareDataPage(MainPage):
-    def __init__(self, dash_app):
+    def __init__(self, dash_app,
+                 page_name: str = 'Prepare Data',
+                 page_id: str = 'prepare-data',
+                 url: str = 'prepare-data'):
         super().__init__(dash_app,
-                         page_name='Prepare Data',
-                         page_id='prepare-data',
-                         url='prepare-data',
+                         page_name=page_name,
+                         page_id=page_id,
+                         url=url,
                          )
 
     def get_layout(self):
-        input_tables = self.dash_app.get_input_table_names()
+        # input_tables = self.dash_app.get_input_table_names()
         layout = html.Div([
-
-            dbc.Card([
-                dbc.CardHeader('Input Table', style= {'fullscreen':True}),
-                dbc.CardBody(
-                    dcc.Dropdown(id='input_table_drpdwn',
-                                 options=[ {'label': i, 'value': i}
-                                           for i in input_tables],
-                                 value=input_tables[0],
-                                 style = {'width': '75vw','height':'2vw'},
-                                 ),
-                ),
-            ], style = {'width': '80vw'}),
+            self.get_input_table_selection_card(),
+            # dbc.Card([
+            #     dbc.CardHeader('Input Table', style= {'fullscreen':True}),
+            #     dbc.CardBody(
+            #         dcc.Dropdown(id='input_table_drpdwn',
+            #                      options=[ {'label': i, 'value': i}
+            #                                for i in input_tables],
+            #                      value=input_tables[0],
+            #                      style = {'width': '75vw','height':'2vw'},
+            #                      ),
+            #     ),
+            # ], style = {'width': '80vw'}),
 
             dbc.Card([
                 # dbc.CardHeader('Input Table'),
-                dbc.CardBody(id = 'input_data_table_card',style = {'width': '79vw'} ),
+                dbc.CardBody(id='input_data_table_card', style={'width': '79vw'} ),
                 html.Div(id="input_data_table_div"),
-            ], style = {'width': '80vw'}),
+            ], style={'width': '80vw'}),
 
-            dbc.Card([
-
-                dbc.CardBody(id = 'input_pivot_table_card',style = {'width': '79vw'}),
-                html.Div(id="input_pivot_table_div"),
-            ], style = {'width': '80vw'})
+            # dbc.Card([
+            #
+            #     dbc.CardBody(id = 'input_pivot_table_card',style = {'width': '79vw'}),
+            #     html.Div(id="input_pivot_table_div"),
+            # ], style = {'width': '80vw'})
+            self.get_pivot_table_card(),
 
         ])
         return layout
+
+    def get_input_table_selection_card(self):
+        input_tables = self.dash_app.get_input_table_names()
+        card = dbc.Card([
+            dbc.CardHeader('Input Table', style={'fullscreen': True}),
+            dbc.CardBody(
+                dcc.Dropdown(id='input_table_drpdwn',
+                             options=[ {'label': i, 'value': i}
+                                       for i in input_tables],
+                             value=input_tables[0],
+                             style = {'width': '75vw', 'height': '2vw'},
+                             ),
+            ),
+        ], style={'width': '80vw'})
+        return card
+
+    def get_pivot_table_card(self):
+        card = dbc.Card([
+            dbc.CardBody(id='input_pivot_table_card', style={'width': '79vw'}),
+            html.Div(id="input_pivot_table_div"),
+        ], style={'width': '80vw'})
+        return card
 
 
     def update_data_and_pivot_input_table_callback(self, scenario_name, table_name):