+++ test/fruit_distribution/fruit_dash_app.py
@@ -43,8 +43,8 @@ How-To create a DO Dashboard:
 """
 
 class FruitDashApp(DoDashApp):
-    def __init__(self, db_credentials: Dict, schema: str = None, cache_config: Dict = None,
-                 port: int = 8050, debug: bool = False, host_env: str = None):
+    def __init__(self, db_credentials: Dict, schema: str = None, db_echo: bool = False, cache_config: Dict = None,
+                 port: int = 8050, dash_debug: bool = False, host_env: str = None):
         visualization_pages = [
             KpiPage(self),
             DemandPage(self),
@@ -56,13 +56,14 @@ class FruitDashApp(DoDashApp):
         data_manager_class = FruitDataManager
         plotly_manager_class = FruitPlotlyManager
         super().__init__(db_credentials, schema,
+                         db_echo=db_echo,
                          logo_file_name=logo_file_name,
                          cache_config=cache_config,
                          visualization_pages = visualization_pages,
                          database_manager_class=database_manager_class,
                          data_manager_class=data_manager_class,
                          plotly_manager_class=plotly_manager_class,
-                         port=port, debug=debug, host_env=host_env)
+                         port=port, dash_debug=dash_debug, host_env=host_env)
 
 
     def create_main_pages(self) -> List[MainPage]: