+++ test/fruit_distribution/fruit_index.py
@@ -16,9 +16,10 @@ else:
     from my_secrets.db2wh import DB2Cloud_DO_Dashboards_credentials
     DB2_credentials = DB2Cloud_DO_Dashboards_credentials
 
-DA = FruitDashApp(db_credentials=DB2_credentials, schema='FRUIT_V2', debug=True, host_env=host_env,
-#                   port=8051
-                 )
+DA = FruitDashApp(db_credentials=DB2_credentials, schema='FRUIT_V2', dash_debug=True, host_env=host_env,
+                  #                   port=8051,
+                  db_echo=True,
+                  )